OMG IT WORKS! My panic attack can leave me now
Okay so i found that when I subscribe to an observable in a service is different when i subscribe to the observable in the component.

Initially i thought subscribing to the service in the component would reduce boilerplate, but then i think that means that the component is not listening for changes.